ft-opt: light refactor
