+++ noptr/internal/toptr.go
@@ -1,4 +1,4 @@
-package pkg
+package internal
 
 func ToPtr[T any](v T) *T {
 	return &v
